safari-534.53-branch/LayoutTests/java/lc3/JSNumber/ToInt-004-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion number conversion to int

FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
FAIL java.lang.Integer.MIN_VALUE+1: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE+1: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
FAIL java.lang.Integer.MIN_VALUE +1: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE +1: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
FAIL java.lang.Integer.MIN_VALUE +1: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE +1: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
FAIL java.lang.Integer.MIN_VALUE: ReferenceError: Can't find variable: java
PASS typeof dt.getInteger() is "number"
PASS successfullyParsed is true

TEST COMPLETE
 
